backend/requirements.txt
fastapi
uvicorn
google-generativeai
qdrant-client
pydantic
python-dotenv